+++ src/robot_state.cpp
@@ -54,11 +54,13 @@ std::ostream& operator<<(std::ostream& ostream, const franka::RobotState& robot_
   ostream << "{\"O_T_EE\": " << robot_state.O_T_EE << ", \"O_T_EE_d\": " << robot_state.O_T_EE_d
           << ", \"F_T_EE\": " << robot_state.F_T_EE << ", \"EE_T_K\": " << robot_state.EE_T_K
           << ", \"m_load\": " << robot_state.m_load << ", \"F_x_Cload\": " << robot_state.F_x_Cload
-          << ", \"I_load\": " << robot_state.I_load << ", \"elbow\": " << robot_state.elbow
-          << ", \"elbow_d\": " << robot_state.elbow_d << ", \"tau_J\": " << robot_state.tau_J
-          << ", \"tau_J_d\": " << robot_state.tau_J_d << ", \"dtau_J\": " << robot_state.dtau_J
-          << ", \"q\": " << robot_state.q << ", \"dq\": " << robot_state.dq
-          << ", \"q_d\": " << robot_state.q_d << ", \"dq_d\": " << robot_state.dq_d
+          << ", \"I_load\": " << robot_state.I_load << ", \"m_ee\": " << robot_state.m_ee
+          << ", \"F_x_Cee\": " << robot_state.F_x_Cee << ", \"I_ee\": " << robot_state.I_ee
+          << ", \"elbow\": " << robot_state.elbow << ", \"elbow_d\": " << robot_state.elbow_d
+          << ", \"tau_J\": " << robot_state.tau_J << ", \"tau_J_d\": " << robot_state.tau_J_d
+          << ", \"dtau_J\": " << robot_state.dtau_J << ", \"q\": " << robot_state.q
+          << ", \"dq\": " << robot_state.dq << ", \"q_d\": " << robot_state.q_d
+          << ", \"dq_d\": " << robot_state.dq_d
           << ", \"joint_contact\": " << robot_state.joint_contact
           << ", \"cartesian_contact\": " << robot_state.cartesian_contact
           << ", \"joint_collision\": " << robot_state.joint_collision